Implemented good CMakeLists.txt

+++ include/World/TileMap.hpp
@@ -10,7 +10,7 @@ class TileMap {
     private:
         const int WIDTH = 100;
         const int HEIGHT = 100;
-        int map[100][100];
+        short map[100][100];
         SpriteSheet &sheet;
 
        sf::RenderTexture map_texture;
